# crates/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenLedger API",
    "description": "OpenLedger, an open-source double-entry ledger on PostgreSQL. The write path is one endpoint under the idempotent replay contract of ADR-0013 — the same key with the same body returns the stored result, marked by the `Idempotency-Replayed` header. The read path is five, and three things about it are contract rather than presentation (ADR-0019): every report answers with the `pinned_cursor` it ran at, including when the caller supplied none, and that value is what re-runs the report; every report AMOUNT is an exact-integer decimal STRING, because a total above 2^53 would be silently rounded by a JSON parser while a posting amount, bounded by its own column, stays a number; and an unknown tenant is answered 200-with-nothing rather than 404, because nothing in the schema declares a tenant and inventing the status would mean inventing the registry.",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "0.0.0"
  },
  "paths": {
...
    "/v1/transactions": {
      "post": {
        "tags": [
          "transactions"
        ],
        "summary": "Post a transaction.",
        "description": "The response set below is this endpoint's real one, not a shared error\nenum's: the eight 422 `type`s named are exactly the ones the writer can\nproduce here, and nothing else is documented (spike 021 found both\ncandidate libraries fanning shared enums across statuses their endpoints\ncannot return — the fix is to declare per endpoint, so this project\ndoes).",
        "operationId": "postTransaction",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransactionBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Replayed: this key was already accepted with this same body. The stored result is re-rendered — never a cached response body — and nothing was written (ADR-0013 §2).",
            "headers": {
              "Idempotency-Replayed": {
                "schema": {
                  "type": "boolean"
                },
                "description": "`true`: this response re-renders a previously stored result."
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionCreated"
                }
              }
            }
          },
          "201": {
            "description": "Posted: this call claimed the idempotency key and wrote the transaction atomically.",
            "headers": {
              "Idempotency-Replayed": {
                "schema": {
                  "type": "boolean"
                },
                "description": "`false`: this response is the first for its idempotency key."
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionCreated"
                }
              }
            }
          },
          "400": {
            "description": "The request body is not syntactically valid JSON. `type` is `invalid_request`; `detail` carries the parser's message.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "413": {
            "description": "The request body exceeds the size limit. `type` is `invalid_request`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "415": {
            "description": "The request's `Content-Type` is not `application/json`. `type` is `invalid_request`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "Refused, and nothing was written. The caller must CHANGE the request to escape — which is why this is 422, not 400 or 409 (ADR-0013 §2). `type` is one of: `invalid_request` (a precondition on the body failed, or a field failed to deserialize into its documented type), `idempotency_key_reused` (same key, different body — send a new key or resend the original request unchanged), `account_unknown` (a posting names an account that does not exist or does not hold that currency), `resolve_target_unknown` (`resolves_id` names no transaction on this tenant's book), `resolve_target_not_pending` (only a pending transaction can be resolved — its status never mutates), `reverse_target_unknown` (`reverses_id` names no transaction on this tenant's book), `reverse_target_not_reversible` (only an ordinary posting that is itself neither a resolution nor a reversal can be reversed), `target_already_superseded` (the named target already has its one supersession — resolved or reversed, either fate is final).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "The write failed; nothing was committed. `type` is `internal`, and the caller gets no internals — the operator's log has the error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/v1/transactions/{transaction_id}": {
      "get": {
        "tags": [
          "transactions"
        ],
        "summary": "Read a transaction back.",
        "description": "No cursor and no chart version: a transaction and its entries are\nimmutable (`ck_txn__append_only` and `ck_entries__append_only`, both\n`ENABLE ALWAYS`), so there is nothing for an as-of to pin (ADR-0019).",
        "operationId": "getTransaction",
        "parameters": [
          {
            "name": "transaction_id",
            "in": "path",
            "description": "The transaction to read.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "tenant_id",
            "in": "query",
            "description": "The book to read.",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "t1"
          }
        ],
        "responses": {
          "200": {
            "description": "The transaction and its entries. `entries` is empty for a void — the zero-posting marker a reversal of a PENDING transaction writes (ADR-0016).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionRead"
                }
              }
            }
          },
          "400": {
            "description": "A required query parameter is missing, or a value would not deserialize into its documented type — including a `transaction_id` path segment that is not a UUID. `type` is `invalid_request`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "No such transaction on this tenant's book. `type` is `transaction_unknown`. Note that an unknown TENANT is not 404 but 200-with-nothing on the report routes, and 404 here: there is no tenant registry to consult, so a wrong `tenant_id` is indistinguishable from a book that does not hold this transaction.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "Refused. `type` is `tenant_mismatch` — the `tenant_id` asked about is not the scope the read path set on the session. Vacuous today by construction and declared anyway (ADR-0019).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "The read failed. `type` is `internal`, and the caller gets no internals — the operator's log has the error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "503": {
            "description": "The read exceeded the read pool's `statement_timeout` (`57014`). `type` is `report_timed_out`. **503, not 500 and not 504**: the service is healthy, the request was too expensive, and retrying it unchanged fails identically (ADR-0019).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "StatusBody": {
        "type": "string",
        "description": "Whether the transaction lands on the books or records money that MAY\nmove. Status never mutates: a pending transaction becomes posted by a NEW\ntransaction naming it in `resolves_id`, never by an update (ADR-0016).",
        "enum": [
          "pending",
          "posted"
        ]
      },
      "TransactionBody": {
        "type": "object",
        "description": "The body of `POST /v1/transactions`.",
        "required": [
          "tenant_id",
          "idempotency_key"
        ],
        "properties": {
          "effective_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "description": "When the movement is deemed to have happened, RFC 3339. Required —\nthe writer will not invent a date — EXCEPT on a reversal, where\nomission means \"the target's own effective_at\" (ADR-0016's soft\nconvention). A supplied date is taken as given, including one below\nthe target's; the cost of that window is recorded in ADR-0016."
          },
          "idempotency_key": {
            "type": "string",
            "description": "Caller-supplied replay key, unique per tenant. Sending the same key\nwith the same body replays the stored result; with a different body it\nis refused as `idempotency_key_reused` (ADR-0013 §2).",
            "example": "charge-1",
            "minLength": 1
          },
          "postings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PostingBody"
            },
            "description": "At least one posting — except on a reversal, which must carry NONE:\nthe server derives the mirror from the target."
          },
          "resolves_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid",
            "description": "The PENDING transaction this one resolves — pending → posted is this\nnew, posted transaction, never an update to the original (ADR-0016).\nThe resolution's postings need not mirror the pending amounts (a\npartial capture resolves with less); a resolving transaction cannot\nitself be pending."
          },
          "reverses_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid",
            "description": "The transaction this one reverses — operational undo, as a NEW\nposted transaction (ADR-0016): a posted target is mirrored in full\n(same legs, directions flipped, derived by the server — send NO\npostings), and a pending target is voided by a zero-posting marker.\nThe target must be an ordinary posting on this tenant's book,\nitself neither a resolution nor a reversal, and not already\nsuperseded."
          },
          "status": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/StatusBody",
                "description": "Omitted means `posted`."
              }
            ]
          },
          "tenant_id": {
            "type": "string",
            "description": "The book this transaction belongs to. Named in the body by decision\n(ADR-0017): data scoping, never an identity claim — the trust story is\nthe deployment perimeter's, and authenticating callers is the\ndeployer's layer.",
            "example": "t1"
          }
        }
      },
...
    }
  },
...
}